EVA-7 Add Ballot backend feature to ballotController.js
